+++ CHANGELOG.rst
@@ -2,6 +2,14 @@
 Changelog for package move_basic
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+0.2.2 (2017-10-08)
+------------------
+* Will reverse for small distances without rotating
+* Try to operate in the frame the goal is in if no map frame
+* Add map_frame param
+* Add frane to points in path
+* Contributors: Jim Vaughan, Rohan Agrawal
+
 0.2.1 (2017-07-17)
 ------------------
 * Add map_frame param